Move devserver to internal/devserver
devserver (the embedded Temporal dev-server launcher) has no consumer outside
the root module. Application code that isn't part of a public API belongs
under internal/, where the import boundary is compiler-enforced. Move it and
update the two importers.

Testing: `go build ./...` passes; the internal/devserver unit tests
(TestAllocatePorts, TestDefaultOutputDirUsesRepoRoot, TestWriteDynamicConfig,
...) pass.

stack-info: PR: https://github.com/temporalio/omes/pull/374, branch: lilydoar/stack/8
