+++ scenarios/project/project_test.go
@@ -11,7 +11,7 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/temporalio/features/sdkbuild"
 	"github.com/temporalio/omes/clioptions"
-	"github.com/temporalio/omes/devserver"
+	"github.com/temporalio/omes/internal/devserver"
 	"github.com/temporalio/omes/internal/workerctl"
 	"github.com/temporalio/omes/loadgen"
 	"github.com/temporalio/omes/versions"
